Fix start ow server helper
At some point, due to unknown cause, owserver started to alter its
behaviour, i.e. it starts and then stop immediately without any error,
instead of releasing the console and continuing in the background. It
works perfectly fine though when you start it in foreground mode. This
fix starts owserver in the foreground and puts it in background
afterwards.